CapsLock++_including_letter-killing-Thunderbird.ahk - Options to kill all letters  or map all letters to lowercase
 intercepts all or almost all keystrokes, transforming them as follows:

 lowercase_lock:  forces all letters to lowercase,  the opposite of CapsLock

kill_letters:  swallows, i.e. disables, all or most unmodified letters.

kill_letters was first used for Thunderbird:  in the main Thunderbird window  there are many keyboard shortcuts that begin with unmodified ordinary printable letters.  this meant that if I was working in this window and a Dragon speech command was misrecognized  as ordinary dictation, bad things would happen. The kill_letters behavior is only implemented  for the Thunderbird main window here.

 these functions are controlled by overloading CapsLock, with various modifiers.  e.g. ^+CapsLock => Lowercase_lock, ^CapsLock sets  lowercase_lock, kill_letters, CapsLock, and some other keys to normal state. Unmodified CapsLock  is disabled, but displays a message box to be less confusing

TBD: as of 2021-07-16  I have encountered a similar problem in GNU  EMACS'  Gnus mail reader,  which also uses many unmodified key mappings. I am quite likely to evolve this script would be useful  in the  Gnus mail reader, and possibly other situations.  So please do not expect stability in the key-bindings